Add few more checks, move common function to separate file
