+++ src/soter/openssl/soter_ec_key_utils.h
@@ -19,7 +19,6 @@
 
 #include <string.h>
 
-#include <openssl/bn.h>
 #include <openssl/obj_mac.h>
 
 #include "soter/soter_ec_key.h"
@@ -95,32 +94,5 @@ static char* ec_priv_key_tag(int curve)
         return NULL;
     }
 }
-
-#if OPENSSL_VERSION_NUMBER < 0x10100000L
-/* Simple implementation for OpenSSL <1.1.0 where this function is missing */
-static int BN_bn2binpad(const BIGNUM* a, unsigned char* to, int tolen)
-{
-    int bn_size = BN_num_bytes(a);
-    int bytes_copied;
-
-    if (a == NULL || to == NULL) {
-        return -1;
-    }
-
-    if (tolen < bn_size) {
-        return -1;
-    }
-
-    bytes_copied = BN_bn2bin(a, to + (tolen - bn_size));
-
-    if (bytes_copied != bn_size) {
-        return -1;
-    }
-
-    memset(to, 0, (size_t)(tolen - bn_size));
-
-    return tolen;
-}
-#endif
 
 #endif /* THEMIS_SOTER_EC_KEY_UTILS_H */
